will FIX: few minor refactoring and updates before making PR
