+++ solar-panel-planner/src/Layout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom"
-import Navbar from "./components/Navbar"
+import Navbar from "./ui/Navbar"
 
 export default function Layout() {
     return (
