+++ solar-panel-planner/src/ui/Footer.jsx
