+++ solar-panel-planner/src/ui/Navbar.jsx
